Rename data formatter files.
